+++ packages/flutter/lib/scheduler.dart
@@ -11,5 +11,6 @@
 /// For example, an idle-task is only executed when no animation is running.
 library scheduler;
 
-export 'src/scheduler/scheduler.dart';
+export 'src/scheduler/binding.dart';
+export 'src/scheduler/priority.dart';
 export 'src/scheduler/ticker.dart';
